fix : IDE-8 lint 설정 경고 해결
