+++ src/pages/IDEPage/FileTree/FileTree.tsx
@@ -3,29 +3,7 @@ import { Tree } from 'react-arborist';
 import { Node } from './Node';
 import CreateFile from './CreateFile';
 import { Resizable } from 're-resizable';
-
-export const data = [
-  {
-    id: '1',
-    name: 'public',
-    children: [{ id: 'c1-1', name: 'index.html' }],
-  },
-  {
-    id: '2',
-    name: 'src',
-    children: [
-      {
-        id: 'c2-1',
-        name: 'App.js',
-        children: [{ id: '44', name: 'sdds.html' }],
-      },
-      { id: 'c2-2', name: 'index.js' },
-      { id: 'c2-3', name: 'styles.css' },
-    ],
-  },
-  { id: '3', name: 'package.json' },
-  { id: '4', name: 'README.md' },
-];
+import { data } from '../../../constants/tempFileTreeData';
 
 const FileTree = () => {
   return (
