+++ src/pages/IDEPage/Terminal/Terminal.tsx
@@ -90,7 +90,7 @@ const Terminal = () => {
 
       deactivateClient();
     };
-  }, []);
+  });
 
   return (
     <TerminalContainer>
